+++ data-features/weight-history/src/commonMain/kotlin/com/grippo/data.features.weight.history/domain/WeightHistoryRepository.kt
@@ -8,6 +8,6 @@ internal interface WeightHistoryRepository {
     fun observeLastWeight(): Flow<WeightHistory?>
 
     suspend fun getWeightHistory(): Result<Unit>
-    suspend fun updateWeight(value: Double): Result<Unit>
+    suspend fun updateWeight(value: Float): Result<Unit>
     suspend fun removeWeight(id: String): Result<Unit>
 }
